[Engine.Tests] Fix compilation of Engine.Tests project.

+++ sources/engine/SiliconStudio.Paradox.Engine.Tests/TestAnimationChannel.cs
@@ -26,7 +26,7 @@ namespace SiliconStudio.Paradox.Engine.Tests
                     if (x.Ticks == 196588)
                     {
                     }
-                    return (float)Math.Sin((double)x.Ticks / (double)CompressedTimeSpan.FromSeconds(10.0).Ticks * Math.PI * 2.0);
+                    return (float)Math.Sin(x.Ticks / (double)CompressedTimeSpan.FromSeconds(10.0).Ticks * Math.PI * 2.0);
                 };
             animationChannel.Fitting(
                 curve,
